+++ src/App.jsx
@@ -1,18 +1,14 @@
 import { Provider } from 'react-redux';
-import RegisterPage from './pages/RegisterPage/RegisterPage';
-import { AppContainer, PageWrapper } from './styles/GlobalStyle';
 import { store } from './redux/store';
 import Router from './shared/Router';
+import { AppContainer } from './styles/GlobalStyle/AppStyle';
 
 const App = () => {
   return (
     <Provider store={store}>
-      <PageWrapper>
-        <AppContainer>
-          <Router />
-          <RegisterPage />
-        </AppContainer>
-      </PageWrapper>
+      <AppContainer>
+        <Router />
+      </AppContainer>
     </Provider>
   );
 };
