+++ src/components/Common/LoadingBar.jsx
@@ -1,4 +1,4 @@
-import { LoadingWrapper, Ellipsis1, Ellipsis2, Ellipsis3, Ellipsis4 } from '../styles/LoadingBarStyle'; // 스타일을 임포트합니다.
+import { LoadingWrapper, Ellipsis1, Ellipsis2, Ellipsis3, Ellipsis4 } from '../../styles/Common/LoadingBarStyle'; // 스타일을 임포트합니다.
 
 const LoadingBar = () => {
   return (
