+++ src/main.jsx
@@ -1,11 +1,12 @@
 import ReactDOM from 'react-dom/client';
 import App from './App.jsx';
-import './index.css';
 import { Provider } from 'react-redux';
 import store from './redux/store.js';
+import { GlobalStyle } from './styles/GlobalStyle/GlobalStyle.js';
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <Provider store={store}>
+    <GlobalStyle />
     <App />
   </Provider>
 );
