+++ src/pages/DetailPage/Detail.jsx
