+++ src/pages/RegisterPage/RegisterPage.jsx
@@ -1,5 +1,5 @@
-import styled from 'styled-components';
 import RegisterForm from '../../components/RegisterComponents/RegisterForm';
+import { StRegisterPageWrapper } from '../../styles/Register/RegisterStyle';
 
 function RegisterPage() {
   return (
@@ -10,17 +10,3 @@ function RegisterPage() {
 }
 
 export default RegisterPage;
-
-const StRegisterPageWrapper = styled.div`
-  border: 1px solid black;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  position: absolute;
-  width: 60vw;
-  height: 80vh;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  text-align: center;
-`;
