+++ src/styles/Common/LoadingBarStyle.js
