+++ src/styles/GlobalStyle/AppStyle.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const AppContainer = styled.div`
+  text-align: center;
+  max-width: 1400px;
+  min-width: 960px;
+  margin: 0 auto;
+  padding-top: 70px;
+`;
