+++ src/styles/Register/RegisterStyle.js
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+export const StRegisterPageWrapper = styled.div`
+  border: 1px solid black;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  position: absolute;
+  width: 60vw;
+  height: 80vh;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  text-align: center;
+`;
